FIX: 리뷰 반영 (Category, Store 데이터 파일 분리)
